refactor(app): split terminal serialization helpers
